reportme_api/requirements.txt
asgiref==3.9.2
cx_Oracle==8.3.0
Django==5.2.6
django-cors-headers==4.9.0
django-filter==23.5
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
drf-spectacular==0.27.2
et_xmlfile==2.0.0
openpyxl==3.1.5
psycopg2-binary==2.9.10
PyJWT==2.10.1
pymssql==2.3.7
python-decouple==3.8
sqlparse==0.5.3
pyodbc==5.0.1
